feat: 상속을 통해 ViewController, ViewModel 의 공통기능 구현
- cell 의 image 의 button 이 나타나는 오류 수정
