+++ OpenMarket/OpenMarket/Scene/Managing/Register/ViewModel/RegisterViewModel.swift
@@ -4,31 +4,16 @@
 //
 //  Created by 박세리 on 2022/05/24.
 //
-
 import UIKit
 
-final class RegisterViewModel {
-    enum Section: CaseIterable {
-        case main
-    }
-    
-    typealias DataSource = UICollectionViewDiffableDataSource<Section, ImageInfo>
-    typealias Snapshot = NSDiffableDataSourceSnapshot<Section, ImageInfo>
-    
-    var datasource: DataSource?
-    
-    private let productsAPIServie = APIProvider<Products>()
-    private(set) var images: [ImageInfo] = []
-    
-    weak var delegate: EditAlertDelegate?
-    
-    func requestPost(_ productsPost: ProductsPost) {
+final class RegisterViewModel: ManagingViewModel{
+    func requestPost(_ productsPost: ProductsPost, completion: @escaping () -> ()) {
         let endpoint = EndPointStorage.productsPost(productsPost)
         
         productsAPIServie.registerProduct(with: endpoint) { [weak self] result in
             switch result {
-            case .success(let data):
-                print(data)
+            case .success():
+                completion()
             case .failure(let error):
                 DispatchQueue.main.async {
                     self?.delegate?.showAlertRequestError(with: error)
@@ -51,18 +36,5 @@ final class RegisterViewModel {
     
     func removeLastImage() {
         images.removeLast()
-    }
-    
-    private func generateUUID() -> String {
-        return UUID().uuidString + ".jpg"
-    }
-    
-    private func applySnapshot() {
-        DispatchQueue.main.async {
-            var snapshot = Snapshot()
-            snapshot.appendSections(Section.allCases)
-            snapshot.appendItems(self.images, toSection: .main)
-            self.datasource?.apply(snapshot)
-        }
     }
 }
